LRCI-3544 Change around the packages for dalo & repository objects
