+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/project/dalo/ProjectDALO.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.jethr0.build;
+package com.liferay.jethr0.project.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
+import com.liferay.jethr0.project.Project;
+import com.liferay.jethr0.project.ProjectFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -24,14 +26,14 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class BuildDALO extends BaseEntityDALO<Build> {
+public class ProjectDALO extends BaseEntityDALO<Project> {
 
 	@Override
-	protected EntityFactory<Build> getEntityFactory() {
-		return _buildFactory;
+	protected EntityFactory<Project> getEntityFactory() {
+		return _projectFactory;
 	}
 
 	@Autowired
-	private BuildFactory _buildFactory;
+	private ProjectFactory _projectFactory;
 
 }
