+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/project/dalo/ProjectPrioritizerDALO.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.jethr0.project;
+package com.liferay.jethr0.project.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
+import com.liferay.jethr0.project.prioritizer.ProjectPrioritizer;
+import com.liferay.jethr0.project.prioritizer.ProjectPrioritizerFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -24,14 +26,14 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectDALO extends BaseEntityDALO<Project> {
+public class ProjectPrioritizerDALO extends BaseEntityDALO<ProjectPrioritizer> {
 
 	@Override
-	protected EntityFactory<Project> getEntityFactory() {
-		return _projectFactory;
+	protected EntityFactory<ProjectPrioritizer> getEntityFactory() {
+		return _projectPrioritizerFactory;
 	}
 
 	@Autowired
-	private ProjectFactory _projectFactory;
+	private ProjectPrioritizerFactory _projectPrioritizerFactory;
 
 }
