+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/project/repository/ProjectRepository.java
@@ -12,15 +12,17 @@
  * details.
  */
 
-package com.liferay.jethr0.project;
+package com.liferay.jethr0.project.repository;
 
 import com.liferay.jethr0.build.Build;
-import com.liferay.jethr0.dalo.ProjectToBuildsDALO;
 import com.liferay.jethr0.dalo.ProjectToTasksDALO;
 import com.liferay.jethr0.dalo.ProjectsToGitBranchesDALO;
 import com.liferay.jethr0.dalo.ProjectsToTestSuitesDALO;
 import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.gitbranch.GitBranch;
+import com.liferay.jethr0.project.Project;
+import com.liferay.jethr0.project.dalo.ProjectDALO;
+import com.liferay.jethr0.project.dalo.ProjectToBuildsDALO;
 import com.liferay.jethr0.task.Task;
 import com.liferay.jethr0.testsuite.TestSuite;
